api/requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==4.0.4
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-exceptions-hog==0.2.0
idna==3.3
packaging==21.3
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
wrapt==1.14.0
